backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0.post1
numpy==1.24.3
pandas==2.1.1
scikit-learn==1.3.0
python-dotenv==1.0.0
httpx==0.26.0
beautifulsoup4==4.12.2
python-multipart==0.0.6
psycopg2-binary==2.9.9
fake-useragent==1.4.0
pytest==7.4.3
pytest-asyncio==0.21.1
tenacity==8.2.3
joblib==1.3.2
pydantic[email]==2.5.2
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
matplotlib==3.8.0
seaborn==0.13.0
apscheduler==3.10.4